move JWT check to `resolveAccessToken` to make it work with realtime
